requirements.txt
python-telegram-bot==21.9
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
psycopg2-binary
requests
Pillow
gspread
oauth2client
pytz
pytz
